+++ pkgs/test/test/runner/compiler_runtime_matrix_test.dart
@@ -22,8 +22,7 @@ void main() {
   for (var runtime in Runtime.builtIn) {
     for (var compiler in runtime.supportedCompilers) {
       // Ignore the platforms we can't run on this OS.
-      if (((runtime == Runtime.internetExplorer || runtime == Runtime.edge) &&
-              !Platform.isWindows) ||
+      if ((runtime == Runtime.edge && !Platform.isWindows) ||
           (runtime == Runtime.safari && !Platform.isMacOS)) {
         continue;
       }
@@ -116,8 +115,7 @@ void main() {
       },
           skip: compiler == Compiler.dart2wasm
               ? 'Wasm tests are experimental and require special setup'
-              : [Runtime.firefox, Runtime.nodeJS, Runtime.internetExplorer]
-                          .contains(runtime) &&
+              : [Runtime.firefox, Runtime.nodeJS].contains(runtime) &&
                       Platform.isWindows
                   ? 'https://github.com/dart-lang/test/issues/1942'
                   : null);