+++ pkgs/test/test/runner/runner_test.dart
@@ -121,14 +121,13 @@ Output:
 
 final _runtimes = '[vm (default), chrome, firefox'
     '${Platform.isMacOS ? ', safari' : ''}'
-    '${Platform.isWindows ? ', ie' : ''}, edge, node]';
+    ', edge, node]';
 
 final _runtimeCompilers = [
   '[vm]: kernel (default), source, exe',
   '[chrome]: dart2js (default), dart2wasm',
   '[firefox]: dart2js (default), dart2wasm',
   if (Platform.isMacOS) '[safari]: dart2js (default)',
-  if (Platform.isWindows) '[ie]: dart2js (default)',
   '[edge]: dart2js (default)',
   '[node]: dart2js (default)',
 ].map((str) => '                                      $str').join('\n');